requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.3.2
pytesseract==0.3.8
gunicorn==20.1.0
boto3==1.20.0
python-dotenv==0.19.0
